ent-v4/calender-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
cassandra-driver==3.25.0
minio==7.1.14
httpx==0.24.1
python-dotenv==1.0.0
requests==2.29.0